fix: stop esmfold2's autocast triangle promoting its operands to float32
`_autocast_triangle` multiplied a bfloat16 `routed` by a float32
`pair_mask` (model.py:1310) with no cast on the mask, then cast the
product to float32 explicitly. Either line alone promotes
[batch, N, N, 2*c_z] -- the widest tensor the block owns -- on the way
into a contraction that narrows its operands back to bfloat16 two lines
later, so the float32 width was created and never used. Both are gone.
The float32 accumulation is unchanged, kept by the einsum's existing
preferred_element_type.

Bit-identical, not merely exact arithmetic: float32 represents every
bfloat16 value, so the round-trip was the identity, and pair_mask is
0.0/1.0, so the multiply is exact at either width. The test asserts
bitwise equality and separately shows a non-0/1 mask would break it.

This is 63dd96e's third arena fix, which never reached the released
path. That commit narrowed three tenants and measured 23.47 -> 21.08 GiB
at 1,003 tokens; e1316e2 then added the native-autocast redirect and only
two followed it -- _autocast_transition carries its own chunk loop and
outer_product_mean blocks on both arms, but the dtype fix stayed in
triangle_multiplicative's body, which the released default never reaches.
test_triangle_multiplication_keeps_its_operand_width missed it for the
same reason: it called triangle_multiplicative without native_autocast.
The new test names the branch the default takes.

Realised-dtype trace of one pair_update_block, two triangle engines, at
2*c_z: convert_element_type 4 -> 2 and mul 2 -> 0. One float32 tenant of
that width survives per engine, the sigmoid's operand, which is
load-bearing. Flipping each line back independently fails the new test
with a different census: the mask line adds a `mul`, the astype line a
second `convert_element_type`.

Also corrects the three places that called this port's peak a
`num_samples * L^2 * 4*c_z` arena (model.py, docs/benchmark.md,
docs/cli.md). The trunk has no sample axis: run_loops and folding_trunk
return [batch, tokens, tokens, d_pair] and the sample factor first
multiplies the batch at model.py:1157, downstream of both.
model.py:196-204 already said so twenty lines below the wrong formula.

+++ src/foldjax/models/esmfold2/models/trunk.py
@@ -120,11 +120,22 @@ def _autocast_triangle(pair, params, prefix, outgoing, mask, eps):
     gate = jax.nn.sigmoid(logits.astype(jnp.float32)).astype(jnp.bfloat16)
     routed = signal * gate
     if mask is not None:
-        routed = routed * mask[..., None]
-    left, right = jnp.split(routed.astype(jnp.float32), 2, axis=-1)
+        # Cast down, the way `triangle_multiplicative`'s own body does and the
+        # way upstream does at `modeling_esmfold2.py:1098` -- "so masking does
+        # not promote the O(N^3) contraction to fp32". `pair_mask` is float32
+        # (`model.py:1310`) and `routed` is bfloat16, so without this the
+        # multiply promotes the widest tensor this block owns.
+        routed = routed * mask[..., None].astype(routed.dtype)
+    left, right = jnp.split(routed, 2, axis=-1)
     equation = "bikd,bjkd->bijd" if outgoing else "bkid,bkjd->bijd"
-    # Pinned native default chunks the output i dimension at 64; autocast
-    # narrows these explicit float() operands and stores the contraction BF16.
+    # Pinned native default chunks the output i dimension at 64. Upstream
+    # spells `routed.float().chunk(2, ...)` here and lets autocast narrow the
+    # operands back at the GEMM; this reproduces the arithmetic without
+    # materialising the float32 in between. Bit-identical, not merely exact:
+    # bfloat16 -> float32 -> bfloat16 is the identity, and `pair_mask` is
+    # 0.0/1.0, so neither step the promotion used to add could change a value.
+    # The float32 accumulation is kept by `preferred_element_type` below, as
+    # it already was.
     chunks = []
     for start in range(0, pair.shape[1], 64):
         operand = (
